itR-16: refactor: add spinner
