+++ src/components/LoadingSpinner/LoadingSpinner.css
@@ -8,10 +8,17 @@
 }
 
 .loading-spinner {
+  position: absolute;
+  top: 110px;
+  left: 225px;
   width: 50px;
   height: 50px;
-  border: 10px solid #f3f3f3;
-  border-top: 10px solid #383636;
+  border: 10px solid #e2e3d1;
+  border-top: 10px solid #c7a51c;
   border-radius: 50%;
   animation: spinner 1.5s linear infinite;
+}
+
+.spinner-container {
+  position: relative;
 }
