+++ src/components/LoadingSpinner/LoadingSpinner.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import "./LoadingSpinner.css";
 
 function LoadingSpinner() {
