+++ src/components/ModalWindow/ModalWindow.tsx
@@ -1,15 +1,36 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import "./ModalWindow.css";
 import { ReactComponent as Cross } from "../../assets/cross.svg";
 import { useIntl } from "react-intl";
 import RepositoriesList from "../RepositoriesList/RepositoriesList";
+import LoadingSpinner from "../LoadingSpinner/LoadingSpinner";
+import axios from "axios";
 
 interface ModalWindowProps {
   setModalActive: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
+type SearchResult = Array<Repository>;
+
+type Repository = {
+  name: string;
+  id: number;
+  html_url: string;
+};
+
 function ModalWindow(props: ModalWindowProps) {
   const intl = useIntl();
+  const [isLoading, setIsLoading] = useState<boolean>(true);
+  const [repositories, setRepositories] = useState<Repository[]>([]);
+
+  useEffect(() => {
+    axios
+      .get<SearchResult>("https://api.github.com/users/MilaKolasava/repos")
+      .then((response) => {
+        setRepositories(response.data);
+        setIsLoading(false);
+      });
+  }, []);
 
   return (
     <div className="modalWindow" data-testid="modal-window">
@@ -28,7 +49,14 @@ function ModalWindow(props: ModalWindowProps) {
             </h2>
           </div>
         </div>
-        <RepositoriesList />
+        {isLoading ? (
+          <LoadingSpinner />
+        ) : (
+          <RepositoriesList
+            setIsLoading={setIsLoading}
+            repositories={repositories}
+          />
+        )}
       </div>
     </div>
   );
