+++ src/components/RepositoriesList/RepositoriesList.tsx
@@ -1,8 +1,4 @@
-import axios from "axios";
 import React, { useEffect, useState } from "react";
-import LoadingSpinner from "../LoadingSpinner/LoadingSpinner";
-
-type SearchResult = Array<Repository>;
 
 type Repository = {
   name: string;
@@ -10,11 +6,14 @@ type Repository = {
   html_url: string;
 };
 
-function RepositoriesList() {
-  const [repositories, setRepositories] = useState<Repository[]>([]);
+interface RepositoriesListsProps {
+  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>;
+  repositories: Repository[];
+}
+
+function RepositoriesList(props: RepositoriesListsProps) {
   const [selectedRepositiry, setSelectedRepository] =
     useState<Repository | null>(null);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
 
   useEffect(() => {
     if (selectedRepositiry) {
@@ -22,20 +21,9 @@ function RepositoriesList() {
     }
   }, [selectedRepositiry]);
 
-  useEffect(() => {
-    axios
-      .get<SearchResult>("https://api.github.com/users/MilaKolasava/repos")
-      .then((response) => {
-        setRepositories(response.data);
-        setIsLoading(false);
-      });
-  }, []);
-
-  return { isLoading } ? (
-    <LoadingSpinner />
-  ) : (
+  return (
     <ol data-testid="repositories-list-items">
-      {repositories.map((repository) => (
+      {props.repositories.map((repository) => (
         <li
           key={repository.id}
           className={selectedRepositiry === repository ? "selected" : ""}
