+++ src/components/Weather/Weather.css
@@ -1,6 +1,6 @@
 .weather-wrapper {
   width: 320px;
-  height: 200px;
+  height: 185px;
   position: absolute;
   bottom: 31px;
   right: 0;
